refactor: extract web flow code (android + iOS)
